Split code to one type per file
