+++ src/EditorFeatures/Core/IIntelliSensePresenterSession.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System;
-using Microsoft.VisualStudio.Text;
-using Microsoft.VisualStudio.Text.Editor;
 
 namespace Microsoft.CodeAnalysis.Editor
 {
@@ -10,17 +8,5 @@ namespace Microsoft.CodeAnalysis.Editor
     {
         void Dismiss();
         event EventHandler<EventArgs> Dismissed;
-    }
-
-    internal interface IIntelliSensePresenter<TPresenter>
-        where TPresenter : IIntelliSensePresenterSession
-    {
-        TPresenter CreateSession(ITextView textView, ITextBuffer subjectBuffer);
-    }
-
-    internal interface IIntelliSensePresenter<TPresenter, TEditorSession> : IIntelliSensePresenter<TPresenter>
-        where TPresenter : IIntelliSensePresenterSession
-    {
-        TPresenter CreateSession(ITextView textView, ITextBuffer subjectBuffer, TEditorSession session);
     }
 }
