+++ src/EditorFeatures/Core/IIntelliSensePresenter`2.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+using Microsoft.VisualStudio.Text;
+using Microsoft.VisualStudio.Text.Editor;
+
+namespace Microsoft.CodeAnalysis.Editor
+{
+    internal interface IIntelliSensePresenter<TPresenter, TEditorSession> : IIntelliSensePresenter<TPresenter>
+        where TPresenter : IIntelliSensePresenterSession
+    {
+        TPresenter CreateSession(ITextView textView, ITextBuffer subjectBuffer, TEditorSession session);
+    }
+}
